fix(parser): resolve IfcComplexProperty's nested HasProperties in the on-demand STEP path
The on-demand property-value parser (extractPropertiesOnDemand /
extractTypePropertiesOnDemand / the material-pset resolver in
on-demand-extractors.ts) had no case for IfcComplexProperty, whose EXPRESS
attributes are [Name, Description, UsageName, HasProperties]. It fell
through to the IfcPropertySingleValue default branch, which reads attribute
index 2 as NominalValue — for IfcComplexProperty that slot is UsageName, a
label. So a complex property showed its UsageName as the value and every
nested property in HasProperties silently vanished from the panel/query
output, with no error.

resolveComplexPropertyValue (new, packages/parser/src/property-value-parser.ts)
walks HasProperties, recursing into nested IfcComplexProperty, and produces
a "Name: value, ..." display plus a flat `values` candidate list, matching
the existing enumerated/list/bounded/table-value shape.
parsePropertyValueWithComplex dispatches to it for IfcComplexProperty and to
the existing single-entity parser otherwise; columnar-parser.ts's
extractPropertiesOnDemand and on-demand-extractors.ts's pset/material-pset
resolvers now call it instead of the single-entity parser directly.

parsePropertyValue/extractNumericValue/the new functions were split into
property-value-parser.ts (re-exported from on-demand-extractors.ts) to keep
on-demand-extractors.ts under its module-size budget.

RED: packages/parser/test/on-demand-complex-property.test.ts — an
IfcComplexProperty's nested Width/Height single-values were dropped
(complex.values undefined, complex.value === 'Reference' the UsageName).
GREEN after the fix. Existing same-shape single-value/enum/list/bounded
regression tests (on-demand-properties-regression.test.ts) pass unchanged.
